+++ config/enso/searchable.php
@@ -3,7 +3,6 @@
 use LaravelEnso\Teams\app\Models\Team;
 use LaravelEnso\People\app\Models\Person;
 use LaravelEnso\Companies\app\Models\Company;
-use LaravelEnso\Companies\app\Models\Contact;
 
 return [
     'defaultLabel' => 'name',
@@ -31,14 +30,6 @@ return [
             'attributes' => ['name'],
             'label' => 'name',
             'permissionGroup' => 'administration.teams',
-        ],
-        Contact::class => [
-            'group' => 'Contact',
-            'attributes' => ['person.name', 'person.appellative', 'position'],
-            'label' => 'person.name',
-            'permissions' => ['edit'],
-            'permissionGroup' => 'administration.companies',
-            'routeParam' => ['company' => 'company_id'],
         ],
     ],
 ];